build: :technologist: #58 Add environment variable
Apply env to use,
according to the Next js rules,
in particulary to store information of contact.
A brief explanation of what the environment
variables are and how they are configured is also
added to the Read me file.
Add env.example to.
Finally, based on this change, it is proposed to
translate that file into English.